backend/requirements.txt
# Generated requirements.txt
# Python version: python==3.11.0


annotated-types==0.7.0
asgiref==3.8.1
autocommand==2.2.2
backports.tarfile==1.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
httplib2==0.22.0
idna==3.10
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pip==25.0.1
platformdirs==4.2.2
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pyjwt==2.9.0
pymupdf==1.25.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
tomli==2.0.1
tqdm==4.67.1
typeguard==4.3.0
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.43.0
zipp==3.19.2
dotenv
django
neo4j
celery==5.3.6
redis==5.0.3
flower==1.2.0
sentry-sdk==1.45.0
django-celery-results 
django-celery-beat
boto3 
django-storages
mistralai